Simplify and improve determinism (#12, #28, #29, #31)
Basic HDK is now more like BIP32. ARKG is available for cases
where the instance indeed wants to delegate key generation.
The spec now does not preclude any document authentication
mechanism, so it is applicable to both single-release documents
as well as to for example BBS#. It also does not preclude
delegation to other parties, such as the solution provider.